Refactor action locations
