+++ app/src/app/App.tsx
@@ -1,7 +1,7 @@
 import AppView from 'app/App.View';
 import { AppDispatchProps, AppStateProps } from 'app/AppProps';
 import { connect } from 'react-redux';
-import { getAlgorithms } from 'store/actions/CoreStoreActions';
+import { getAlgorithms } from 'store/actions/AlgorithmsPageActions';
 import { SnowmanDispatch } from 'store/messages';
 import { Store } from 'store/models';
