+++ app/src/pages/AlgorithmsPage/AlgorithmsPage.tsx
@@ -5,7 +5,7 @@ import {
   AlgorithmsPageStateProps,
 } from 'pages/AlgorithmsPage/AlgorithmsPageProps';
 import { connect } from 'react-redux';
-import { getAlgorithms } from 'store/actions/CoreStoreActions';
+import { getAlgorithms } from 'store/actions/AlgorithmsPageActions';
 import { SnowmanDispatch } from 'store/messages';
 import { Store } from 'store/models';
 import { FirstValidMatchingSolution } from 'structs/constants';
