+++ app/src/store/actions/ExperimentsPageActions.ts
@@ -1,11 +1,21 @@
-import { Algorithm } from 'api';
+import { Algorithm, Experiment, ExperimentsApi } from 'api';
 import { DropResult } from 'react-beautiful-dnd';
-import { ExperimentsPageActionTypes } from 'store/actions/actionTypes';
+import {
+  CoreStoreActionTypes,
+  ExperimentsPageActionTypes,
+} from 'store/actions/actionTypes';
+import {
+  SnowmanAction,
+  SnowmanDispatch,
+  SnowmanThunkAction,
+} from 'store/messages';
+import { SUCCESS_TO_DELETE_EXPERIMENT } from 'structs/statusMessages';
 import { getDndDescriptorFromDropResult } from 'utils/dragNDropHelpers';
 import {
   easyPrimitiveAction,
   easyPrimitiveActionReturn,
 } from 'utils/easyActionsFactory';
+import RequestHandler from 'utils/requestHandler';
 
 export const clickOnMatchingSolution = (
   aMatchingSolution: Algorithm
@@ -22,3 +32,33 @@ export const dragNDropAnExperiment = (
     type: ExperimentsPageActionTypes.DRAG_N_DROP_EXPERIMENT,
     payload: getDndDescriptorFromDropResult(aDropResult),
   });
+
+export const getExperiments = (): SnowmanThunkAction<Promise<void>> => async (
+  dispatch: SnowmanDispatch
+): Promise<void> =>
+  RequestHandler(
+    (): Promise<void> =>
+      new ExperimentsApi()
+        .getExperiments()
+        .then(
+          (experiments: Experiment[]): SnowmanAction =>
+            dispatch({
+              type: CoreStoreActionTypes.SET_ALL_EXPERIMENTS,
+              payload: experiments,
+            })
+        )
+        .then(),
+    dispatch
+  );
+
+export const deleteExperiment = (
+  anExperiment: Experiment
+): SnowmanThunkAction<Promise<void>> => async (
+  dispatch: SnowmanDispatch
+): Promise<void> =>
+  RequestHandler(
+    (): Promise<void> =>
+      new ExperimentsApi().deleteExperiment({ experimentId: anExperiment.id }),
+    dispatch,
+    SUCCESS_TO_DELETE_EXPERIMENT
+  ).then((): Promise<void> => dispatch(getExperiments()));
